refactor: move mask matrix related to same script (#118)
